+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Enable azure-scheduled-events for azure workers.
+
 ## [6.3.0] - 2022-02-10
 
 ### Changed
@@ -428,8 +432,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add github release workflows.
 
-[Unreleased]: https://github.com/giantswarm/giantnetes-terraform/compare/v6.3.0...HEAD
-[6.3.0]: https://github.com/giantswarm/giantnetes-terraform/compare/v6.2.2...v6.3.0
+[Unreleased]: https://github.com/giantswarm/giantnetes-terraform/compare/v6.2.2...HEAD
 [6.2.2]: https://github.com/giantswarm/giantnetes-terraform/compare/v6.2.1...v6.2.2
 [6.2.1]: https://github.com/giantswarm/giantnetes-terraform/compare/v6.2.0...v6.2.1
 [6.2.0]: https://github.com/giantswarm/giantnetes-terraform/compare/v6.1.1...v6.2.0